moved downloader and utils to internal package
